Untangle signals from events and improve gameplay
This commit adjusts the functions that respond to our signals,
so that they set bool flags rather than calling further functions.
This moves more of the calling logic to the main game loop.
This stops too much happening while a signal is being processed,
and hopefully fixes a game-breaking bug where the game would lose
track of items in the array representing the snake, which would
cause a crash and/or the death of the player. It should also mean
the game more reliably detects whether the snake has crashed into
itself or not.

A LunchTimer is introduced to control the window of time in which
the snake is immortal after eating. This decouples the snake's
immortality-window from the time it takes other code to run,
removing bugs.

This commit also makes other aspects of the gameplay more fun;
the score is now displayed, and the snake starts in the center
of the screen after death. The deathsound implementation will
need adjusting to match.