Remove needless duplication to position snake
This commit removes lots of stupid stuff we don't need any more.
We no longer make a separate snake_array and match the coordinates
of the segments of the snake to it; we use the segment coordinates
directly. This means we can remove all the code that was there
to support that initial bizarre way of doing things.

We also make segments instantiate offscreen, so the player will not
bump into them or see them flash for a frame. We also set a timer
at the start during which the snake is immortal, to avoid it dying
when it connects its first segment up to its head. We hide this
first segment at the start, too.