Fix collision-detection and remove 'food_flag'
This commit hopefully fixes a game-breaking bug where your
snake ran a risk of suddenly dying if you ran the game on
a slow computer. The method to move the newest segment
forward could end in a crash between the
segment and the head, if the head hung around for
too long. We now disable collision-detection for the newest
segment, so that it can't hit the head by mistake. We then
enable collision-detection for the second-newest segment,
each loop, so that we don't end up with a snake that has
immortal patches.

The .tscns have also been adjusted so that the head is
only looking for crashes in the segments layer, since we
don't use proper collision-detection for the walls anyway.
The food looks for crashes in the head layer.
(and the segment looks for crashes in the head layer, but
does nothing with the information).

This commit also fixes the code for making food for the snake,
so that we can get rid of all traces of the silly 'food_flag'
way of doing things, for good.